Fix the second createDataFrame call site, and a value-type bug it exposed
The ChunkedArray fix only covered ingest_submitted_micro. process_and_publish_macro_batch
has the identical shape: validator.validate(df_submitted) inherits df_submitted's
provenance as one CSV-parsed frame per country concatenated together, which under
pandas 3 backs each string column with a multi-chunk Arrow array that
createDataFrame's pandas path cannot assemble into a RecordBatch. Converted to
records against the already-declared VALIDATED_MACRO_SCHEMA, the same treatment
as the first call site.

That exposed a second, independent failure: OBS_VALUE arrived as a Python int
because pd.to_numeric infers int64 when the round-tripped SDMx-ML serializes a
whole-number observation without a decimal point, and PySpark's verifier checks
exact type rather than coercibility, so DoubleType rejects it. Forced to float
explicitly in load_submission, where the value is parsed.

Verified by running the real ingest-validate-createDataFrame path against a local
Spark session for every arrival in both cycles: createDataFrame succeeds for all
of them, and the only remaining failure is .count() dying on the pre-existing
local JDK/socket issue, unrelated and not present on a Databricks worker.

+++ src/sdmx_rule_validator.py
@@ -320,7 +320,7 @@ class SDMxRuleValidator:
         dim_cols = [dim for dim in self.dimension_order if dim in data.columns]
         data["TIME_SERIES_CODE"] = data[dim_cols].astype(str).agg(".".join, axis=1)
         data = data.rename(columns={"TIME_PERIOD": "DATE"})
-        data["OBS_VALUE"] = pd.to_numeric(data["OBS_VALUE"])
+        data["OBS_VALUE"] = pd.to_numeric(data["OBS_VALUE"]).astype(float)
         data["AGG_CODE"] = AGG_CODE_DEFAULT
         for col in ("OBS_STATUS", "OBS_CONF"):
             if col not in data.columns:
